modify Messages to act as switch between mobile and desktop, start Mobile same as Desktop
